refactor(train): move DLRover API validator to utils
