+++ swift/pipelines/train/sft.py
@@ -38,7 +38,7 @@ class SwiftSft(SwiftPipeline, TunerMixin):
                 from dlrover.trainer.torch.flash_checkpoint.hf_trainer import HfFlashCheckpointer
             except ImportError:
                 raise ValueError('Please install DLRover to use Flash Checkpoint: `pip install dlrover[k8s,torch]`.')
-            from swift.trainers.mixin import _validate_dlrover_flash_checkpoint_api
+            from swift.trainers.utils import _validate_dlrover_flash_checkpoint_api
             _validate_dlrover_flash_checkpoint_api(HfFlashCheckpointer, CheckpointEngine)
 
     def _prepare_generation_config(self):
